unitclothoid.cpp: Rewrite points load/save
Since oPathPoints is defined as a std::vector, dumping TPathPt instances
as Plain Old Data is not valid any more. In fact, C++ compilers would
already warn that offsetof(3) could not be safely used with TPathPt
members since it never was Plain Old Data, but this warning was
suppressed only because raw pointer manipulation had been used instead
of offsetof(3).

On the other hand, dumping binary data is inherently non-portable and
not human-readable, so this commit suggests to dump such data in JSON
format, as cJSON is already used by the project and XML is bad, among
many other things, to store arbitrarily large arrays.


Former-commit-id: 4a447adb452dd9ae85ca4cce9993cbe9ad0487f6
Former-commit-id: b462a614ed53f1938eadb4680a10ef57576b29d9

+++ src/drivers/simplix/src/unitlane.h
@@ -107,10 +107,10 @@ class TLane
 		float	WPitToL;						   //	Lane specfic width to left
 		float	WPitToR;						   //	Lane specfic width to right
 		bool Fix;
-		double MaxSpeed;						   // Max speed	through	this point
-		double AccSpd;						   //	Speed through this point, with modelled	accel
-		double Speed;							   //	Speed through this point (braking only)
-		double FlyHeight;						   // Predicted height	of car above track (flying)
+		float MaxSpeed;						   // Max speed	through	this point
+		float AccSpd;						   //	Speed through this point, with modelled	accel
+		float Speed;							   //	Speed through this point (braking only)
+		float FlyHeight;						   // Predicted height	of car above track (flying)
 
 		double Dist()	const {return DistFromStart;}
 		double WtoL()	const {return WToL;}
